Speed up XXWithScope: queries.
The list of parent/children types returned by moose query was missing subclasses. Thus we needed to check all the time if our class was inheriting from the parentTypes. 

Now we add the subclasses directly in parent/children types method that is cached. This avoid to iterate over the superclasses all the time.