Speed up creation of moose groups.
MooseGroup are long to create because they do a lot of accesses to caches to build a MooseName. 
Now the moose name is stored in a variable since it is something that will always be filled in MooseEntities. 

With this, the creation of a moose group of 411279 entities passed from 2.6sec to 0.4sec.