Add first version of initialize method generation for default value
This version has the advantage of being static and does not rely on values from the slot to work. The counterpart is that it is not possible to modify the metamodel at runtime and to automatically impact all existing instances on which the default value would not have been already written. For a more dynamic behavior, slot handling should be prefered, but it implies the obligation to keep information about if yes or not the property have been accessed/written at least one. Using slots coupled to a dictionnary for simulating instance variables would also mean that we should be able to deal with lazy access to any moose entity property.