Remove support for Trait V1
Traits had 2 implementations in Pharo. After the second version was added, a compatibility package was kept in Pharo to be able to load old code. But users should not use this old way of creating traits. 

Since the new version is present in Pharo since years, I propose to drop the support for the first version that I never saw any usage of in Moose anyway.