Fix bug with class slots
In P12 a bug was fixed about class slots but Moose was exploiting this bug. 

This is a quick fix for it. This code is quite complex but when P12 will be the minimaly maintained Pharo version for Moose we can do some nice simplifications by using the ShiftClassBuilder and not evaluating strings.