Introduce #isRoot boolean property
I want to build a visualization starting with the "root" entities of a model (aka the entities without parent). To make this easy in the query browser I propose to introduce a #isRoot boolean property. 

I introduced it on TEntityMetalevelDependency since this is the trait bringing the concept of parents. While I tired to add tests I found out that we also had some root concept on MooseEntity but I think it is a bad idea to have it there because MooseEntity has no concept of parent. So I removed it.

Fixes #679