Remove MooseGroupTeardownStorage
This storage is never used so Iâ¯propose to remove it. 

Also MooseGroupSetupStorage seems to be used only by the Pharo importer. I want to review some of the caches and bench if it really makes a difference to build the caches at the end of the import only. If the difference is not visible maybe we could flatten the whole storage hierarchy. But I need to do some tests before.