Use a trait for the smalltalk importers utilities
The smalltalk importers sometimes have MooseTask in their inheritance while they are not tasks because we do not have multiple inheritance and we want to share some utilities between importers.

I'm updating this behavior to use a trait instead and now we do not need those extra inheritances that we did not use.