Shared variable properties should be persisted
Smalltalk attributes can be shared variable and we have a way to declare this. 

The problem is that this info can be lost when we flush the caches. This change fix this bug.

I also declared this as a property in the MM so that we can query it in the query browser