Update TAccess to have candidates as targets
Until now we only modelized languages for whom we knew for sure who accessed the variables. But in Python we cannot know. We have candidates like for invocations. 

We discussed with Clotilde and Nicolas and we should probably update TAccess to have candidates. Here is the Famix PR. Before merging we should update VerveineJ.